// frontend/src/models/korisnik.ts
export class Korisnik {
  ime: string;
  prezime: string;
  korisnickoIme: string;
  lozinka: string;
  lozinkaTrajeDo: number;
  mejl: string;
  tip: string;
  status: string;
  telefon: string;
  slika: string;
  nazivOrganizacije: string;
  adresaOrganizacije: string;
  maticniBrojOrganizacije: string;
  radionice: Array<number>;
  skice: Array<string>;
}